// src/movies/domain/dto/movie-domain.dto.ts
export class MovieDomainDto {
  characters: string[];
  created: string;
  director: string;
  edited: string;
  episodeId: number;
  openingCrawl: string;
  planets: string[];
  producer: string;
  releaseDate: string;
  species: string[];
  starships: string[];
  title: string;
  url: string;
  vehicles: string[];
}